+++ js/advertisements.js
@@ -1,0 +1,1 @@
+/*** Модуль "Объявления" ***/
